hackathon/pokedex/requirements.txt
numpy==1.26.4
packaging==23.2
setuptools==69.2.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
requests==2.31.0
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
streamlit==1.32.0 